requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==4.2.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.29
chromadb-client==0.3.30.dev0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.7
dkimpy==1.1.4
dnspython==2.3.0
duckdb==0.8.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.85.1
ffmpy==0.3.0
filelock==3.12.0
filetype==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.6.0
git-filter-repo==2.38.0
google-search-results==2.4.2
gptcache==0.1.18
gradio==3.27.0
gradio_client==0.1.3
grafanalib==0.7.0
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.0
httptools==0.6.0
httpx==0.24.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jq==1.4.1
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.3.1
langchain==0.0.231
langchainplus-sdk==0.0.20
linkify-it-py==2.0.0
llama-index==0.6.23
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
olefile==0.46
onnxruntime==1.15.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.10
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.0
patatt==0.6.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.2
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2-binary==2.9.6
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pymongo==4.4.0
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.12.1
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
reportlab==4.0.4
requests==2.28.2
safetensors==0.3.1
selenium==4.10.0
semantic-version==2.10.0
sentry-sdk==1.28.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.65.0
transformers==4.30.2
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
unstructured==0.8.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
uWSGI==2.0.21
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.2.3
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.1
zipp==3.15.0
zstandard==0.21.0